requirements.txt
httpx
python-dotenv
tavily-python
browser-use-sdk